+++ src/Client/index.ts
@@ -1,4 +1,4 @@
-import { ActivityType, Client, Collection, REST, Routes, ApplicationCommandType, ChatInputCommandInteraction, MessageContextMenuCommandInteraction, UserContextMenuCommandInteraction, GatewayInteractionCreateDispatchData, SnowflakeUtil, CommandInteraction } from "npm:discord.js";
+import { ActivityType, Client, Collection, REST, Routes } from "npm:discord.js";
 import { Command, Config, Event } from "../Interfaces/index.ts";
 import ConfigJson from "../config.json" with { type: "json" };
 import path from "node:path";
@@ -55,22 +55,6 @@ export default class ExtendedClient extends Client {
       console.error(JSON.stringify(e));
     }
 
-    /* Proxy InteractionCreate handler */
-    if (Object.getOwnPropertyDescriptor(this, 'actions')?.value.InteractionCreate !== undefined) {
-      const b = Object.getOwnPropertyDescriptor(this, 'actions')?.value.InteractionCreate;
-      const ohandle = b.handle;
-      const a = new Proxy(ohandle, {
-        apply(target, thisArg, argArray) {
-          if (argArray[0].data.name == "restart") {
-            Deno.writeTextFileSync("/tmp/tijmes-assistant.json", JSON.stringify(argArray));
-          }
-
-          return Reflect.apply(target, thisArg, argArray);
-        }
-      });
-      b.handle = a;
-    }
-
     /* Events */
     const eventPath = path.join(dirname, "..", "Events");
     const events = Deno.readDirSync(eventPath);
@@ -85,59 +69,32 @@ export default class ExtendedClient extends Client {
 
     console.log("Commands and events loaded!");
 
+    /* Proxy InteractionCreate handler */
+    if (
+      Object.getOwnPropertyDescriptor(this, "actions")?.value
+        .InteractionCreate !== undefined
+    ) {
+      const b = Object.getOwnPropertyDescriptor(this, "actions")?.value
+        .InteractionCreate;
+      const ohandle = b.handle;
+      const a = new Proxy(ohandle, {
+        apply(target, thisArg, argArray) {
+          if (argArray[0].data.name == "restart") {
+            Deno.writeTextFileSync(
+              "/tmp/tijmes-assistant.json",
+              JSON.stringify(argArray),
+            );
+          }
+
+          return Reflect.apply(target, thisArg, argArray);
+        },
+      });
+      b.handle = a;
+    }
+
     // Public commands
     await this.login(token);
     console.log("Bot is online!");
-
-    // Edit restart embed
-    try {
-      await Deno.lstat("/tmp/tijmes-assistant.json");
-      const data_j = Deno.readTextFileSync("/tmp/tijmes-assistant.json");
-      Deno.removeSync("/tmp/tijmes-assistant.json");
-
-      let InteractionClass;
-      if (data_j) {
-        const data = JSON.parse(data_j)[0];
-
-        switch (data.data.type) {
-          case ApplicationCommandType.ChatInput:
-            InteractionClass = ChatInputCommandInteraction;
-            break;
-          case ApplicationCommandType.User:
-            InteractionClass = UserContextMenuCommandInteraction;
-            break;
-          case ApplicationCommandType.Message:
-            InteractionClass = MessageContextMenuCommandInteraction;
-            break;
-          default:
-            throw new Error("Invalid interaction");
-        }
-
-        if (InteractionClass !== undefined) {
-          class Interaction extends CommandInteraction {
-            constructor(client: Client<true>, data: GatewayInteractionCreateDispatchData) {
-              super(client, data);
-              this.replied = true;
-            }
-          }
-
-          const interaction = new Interaction(this as Client<true>, data);
-
-          if (SnowflakeUtil.deconstruct(interaction.id).timestamp - BigInt(+new Date()) > -(15 * 1000)) { // interaction too old to edit
-            await interaction.sendEmbed({
-              title: "Restarted!",
-              description: "The bot has succesfully restarted.",
-              ephemeral: true,
-            }, true);
-          }
-        }
-      }
-    } catch (err) {
-      if (!(err instanceof Deno.errors.NotFound)) {
-        console.error(err);
-      }
-    }
-
 
     if (!this.user) return;
     this.user.setPresence({
